Refactor: remove mining-related modules and move cleaning to a stand-alone module
