+++ unpdf/extraction.py
@@ -17,29 +17,7 @@ from tqdm import tqdm
 
 # local packages
 from .entities import DocumentEntity, PageEntity
-
-
-def _preprocess_text(text: str) -> str:
-    """
-    Preprocess a text by removing unprintable characters and standardising to NFD form.
-
-    The function removes unprintable characters , e.g., '\x02', '\x16' but keeps newline and CR characters.
-    NFD normalisation decomposes characters by canonical equivalence, which may be useful for use cases like
-    language identification or machine translation.
-
-    Parameters
-    ----------
-    text : str
-        Input text to be cleaned.
-
-    Returns
-    -------
-    text : str
-        NFD-normalised text with unprintable characters removed.
-    """
-    text = ''.join(char for char in text if char.isprintable() or char in '\r\n')
-    text = normalize('NFD', text)  # '\u00E9' or 'é' (b'\xc3\xa9' in NFC) -> 'é' (b'e\xcc\x81' in NFD)
-    return text
+from .cleaning import _normalise_text
 
 
 def extract_text(
@@ -78,8 +56,8 @@ def extract_text(
     iterable = enumerate(pdf)
     for idx, page in tqdm(iterable, disable=not progress_bar):
         try:
-            text = page.get_textpage().get_text_range(index=0, count=-1)
-            text, error = _preprocess_text(text), False
+            text = page.get_textpage().get_text_range(index=0, count=-1)  # extract text from the whole page
+            text, error = _normalise_text(text), False
         except Exception as e:
             logging.exception(str(e))
             text, error = '', True
